refactor mcp dto normalizers
